feat: add javadoc for packet fields (#2)
* feat: add javadoc for packet fields

* fix: rename command output message fields

* fix: remove non-null annotations from nullable fields

* fix: rename ItemComponentPacket to ItemRegistryPacket

* docs: remove noinspection comments from packet violation javadocs

* chore: format code

* fix: move camera and code builder enums to dedicated packages

* ...

+++ common/src/main/java/org/cloudburstmc/protocol/common/util/VarInts.java
@@ -1,8 +1,6 @@
 package org.cloudburstmc.protocol.common.util;
 
 import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufAllocator;
-import io.netty.buffer.ByteBufUtil;
 import lombok.experimental.UtilityClass;
 
 import java.math.BigInteger;
@@ -14,7 +12,7 @@ public class VarInts {
     private static final BigInteger BIG_INTEGER_80 = BigInteger.valueOf(0x80);
 
     public static void writeInt(ByteBuf buffer, int value) {
-        encode(buffer, ((value << 1) ^ (value >> 31)) & 0xFFFFFFFFL);
+        encode(buffer, (((long) value << 1) ^ (value >> 31)) & 0xFFFFFFFFL);
     }
 
     public static int readInt(ByteBuf buffer) {
